Schema Fields and Their Configuration The schema defines several fields, each corresponding to a property of a user document. Each field is an object that specifies the type of data, whether it's required, and other constraints or default values.
username Field: This field stores the user's username. It is of type String and is both required and unique, meaning every user must have a username, and no two users can share the same username.

email Field: Similar to the username, the email field is of type String, required, and unique. This ensures that every user has an email, and each email is unique across all users.

password Field: This field stores the user's password as a string. It is required but not unique, as different users can have the same password.

profilePicture Field: This field is a string that stores the URL or path to the user's profile picture. It has a default value of an empty string, meaning if a user doesn't set a profile picture, it will default to an empty string.

bio Field: This field stores a short biography of the user as a string. It also has a default value of an empty string.

gender Field: This field is a string that stores the user's gender. It is constrained by an enum (enumeration), meaning it can only be one of the specified values: 'male' or 'female'. This limits the possible values that can be stored in this field.

followers Field: This field stores an array of ObjectIds, each referencing another user document. It represents the users who are following the current user. The ref: 'User' part tells Mongoose that the ObjectIds in this array refer to documents in the User collection.

following Field: Similar to followers, this field stores an array of ObjectIds referencing user documents. It represents the users whom the current user is following.

posts Field: This field stores an array of ObjectIds referencing Post documents. It represents the posts created by the user. The ref: 'Post' indicates that these ObjectIds refer to documents in the Post collection.

bookmarks Field: This field stores an array of ObjectIds referencing Post documents that the user has bookmarked. This allows users to save posts for later viewing.