Socket.IO Connection Handling Connection Event:
Listens for new socket connections.
Retrieves the userId from the connection query parameters.
Maps the userId to the socket ID in userSocketMap.
Update Online Users:

Emits an event (getOnlineUsers) to all connected clients with a list of currently online users. This is done each time a new connection is established.
Disconnect Event:

Listens for the socket disconnect event.
Removes the user ID and socket ID from the userSocketMap.
Emits an updated list of online users to all connected clients to reflect the change.