authSlice Definition name: A string that defines the name of the slice. Here, it is "auth", which is used to identify this slice in the Redux store.
initialState: The initial state for this slice of the Redux store. It contains:

user: Initially set to null. This will store the authenticated user's data.
suggestedUsers: Initially an empty array. This will hold a list of suggested users.
userProfile: Initially null. This will store the profile data of a specific user.
selectedUser: Initially null. This will store the data of the user currently selected (for example, for viewing or messaging).
reducers: An object defining the actions and their respective reducers to update the state. Each key in this object represents an action and the function is the reducer to handle that action:

setAuthUser: Updates the user state with the payload, which represents the authenticated user's data.
setSuggestedUsers: Updates the suggestedUsers state with the payload, which is an array of suggested users.
setUserProfile: Updates the userProfile state with the payload, which represents the profile of a specific user.
setSelectedUser: Updates the selectedUser state with the payload, which is the user currently selected in the application.