postSlice Definition name: A string that identifies this slice of the Redux store. In this case, it's "post", which indicates that this slice manages the state related to posts.
initialState: The initial state of the slice. It defines:

posts: An empty array that will hold the list of posts.
selectedPost: Initially set to null. This will store the currently selected post, if any.
reducers: An object that contains functions for updating the slice's state. Each key corresponds to an action, and the function is a reducer that modifies the state accordingly:

setPosts: A reducer function that updates the posts state with the payload. The payload is expected to be an array of posts.
setSelectedPost: A reducer function that updates the selectedPost state with the payload. The payload represents a single post that is currently selected.