socketSlice Definition name: Specifies the name of the slice. In this case, it's "socketio", indicating that this slice manages the state related to WebSocket or socket connections.
initialState: Sets the initial state for this slice. It contains:

socket: Initially set to null. This is where the WebSocket or socket instance will be stored once it's established.
reducers: Defines the function(s) to handle state updates:

setSocket: A reducer function that updates the socket state with the data provided in the action payload. The payload is expected to be the WebSocket or socket instance.