Export Statements export const { setSocket }: Destructures and exports the action creator for this slice:
setSocket: Action creator function for updating the socket state. This function is used to dispatch actions that set the WebSocket or socket instance in the state.
export default socketSlice.reducer: Exports the reducer function created by createSlice. This reducer manages updates to the socket state in response to actions. It is integrated into the Redux store to handle socket-related state changes.

Summary
createSlice: Simplifies the creation of a Redux slice by combining reducers and actions into one function.
initialState: Defines the initial state of the slice, starting with socket set to null.
reducers: Contains functions for updating the state. In this case, there's only one reducer, setSocket, which updates the socket state.
Action Creator: setSocket is an action creator function that dispatches actions to update the socket state.
Reducer Export: The reducer created by createSlice is exported and integrated into the Redux store to manage socket-related state.
This slice is designed to manage the state of a WebSocket or socket instance in a Redux store. It provides a way to store and update the WebSocket connection instance, which can then be used throughout the application for real-time communication.