rtnSlice Definition name: Specifies the name of the slice. In this case, it's 'realTimeNotification', which indicates that this slice manages notifications in real-time.
initialState: Defines the initial state for the slice. It contains:

likeNotification: An array that will hold notifications related to "likes". Initially, it's an empty array.
reducers: Defines how the state should be updated in response to actions:

setLikeNotification: A reducer function that handles different types of notification updates:
Clearing Notifications: If the action.payload is an empty array, it clears the likeNotification array.
Adding a Like Notification: If the action.payload.type is 'like', it adds the notification to the likeNotification array.
Removing a Dislike Notification: If the action.payload.type is 'dislike', it removes notifications from the likeNotification array where the userId matches the userId in the payload.