ReactDOM Rendering ReactDOM.createRoot:
A new API in React for creating a root to which components are rendered. It’s part of React 18 and onwards for concurrent rendering.
document.getElementById('root'):

The HTML element with the ID root where the React application will be mounted. This element is typically defined in the index.html file.