Imports setPosts:
An action creator from the postSlice Redux slice. This action updates the posts state in the Redux store.
axios:

A promise-based HTTP client for making requests to an API. In this code, it is used to fetch data from the backend.
useEffect:

A React hook used to perform side effects in functional components. In this hook, it's used to execute the fetch operation when the component mounts.
useDispatch:

A React-Redux hook that returns the dispatch function. This function is used to dispatch actions to the Redux store.