Imports setSuggestedUsers:
An action creator from the authSlice Redux slice. This action is used to update the suggestedUsers state in the Redux store with a list of suggested users.
axios:

A promise-based HTTP client used for making HTTP requests. It is used in this code to fetch suggested users from the backend API.
useEffect:

A React hook used for managing side effects in functional components. In this case, it is used to perform the fetch operation when the component mounts.
useDispatch:

A React-Redux hook that returns the dispatch function, which is used to dispatch actions to the Redux store.