Imports setUserProfile:
An action creator from the authSlice Redux slice. It is used to update the userProfile state in the Redux store with the fetched user profile data.
axios:

A promise-based HTTP client used to make HTTP requests. Here, it is employed to send a GET request to fetch the user profile.
useEffect:

A React hook that allows side effects to be managed in functional components. It is used in this code to perform the data fetch operation when the userId changes.
useState:

Although imported, useState is not used in this code. It seems to be a leftover from previous code or future use, as the profile state is managed entirely through Redux.
useDispatch:

A React-Redux hook that returns the dispatch function, used to dispatch actions to the Redux store.