useGetUserProfile Hook Purpose:
The hook fetches the profile information of a user based on the userId and updates the Redux store with this profile data.
Initialization:

The dispatch function is obtained using the useDispatch hook from React-Redux. This function is necessary for dispatching Redux actions.
useEffect Hook:

useEffect is utilized to handle side effects in the component. In this case, it is used to fetch the user profile data when the component mounts or when the userId changes.
fetchUserProfile Function:

An asynchronous function defined inside the useEffect hook. This function handles fetching the user profile data.
axios.get Call:

An HTTP GET request is sent to the API endpoint http://localhost:8000/api/v1/user/${userId}/profile. The URL includes the userId parameter to specify which user's profile to fetch.
The withCredentials: true option is included to ensure that cookies are sent with the request, which may be required for authentication or session management.
Handling Response:

Upon receiving a successful response (res.data.success), the setUserProfile action is dispatched with the retrieved user profile data (res.data.user). This action updates the userProfile state in the Redux store.
Error Handling:

Any errors that occur during the HTTP request are caught and logged to the console.
Dependencies Array:

The userId is included in the dependencies array of useEffect. This means that if the userId changes, the useEffect hook will re-run, triggering a new fetch for the user profile. If the userId remains the same, the fetch operation will not be re-triggered.