Event Handlers:
logoutHandler: Handles user logout. It sends a request to the server to log out, updates the Redux store to clear user data, and navigates the user to the login page. It also shows a success or error message using toast.
sidebarHandler: Handles clicks on sidebar items. Depending on the item clicked, it performs actions like logging out, opening the create post modal, navigating to the user profile, home page, or messages.